Rewrite BsRequest query methods as scopes
These methods where used to query requests that contain reviews that
match the query attributes.
All of this can be done withActiveRecord::QueryMethods and a few
scopes.

This also required a change of how we ordered requests in the staging
controller, since we couldn't use the Array#sort anymore. A nice
side-effect of this is that we now do the request sorting as part of
the SQL query.

Co-authored-by: Dany Marcoux <dmarcoux@suse.com>