[backend] also send snapshot data to notary
Before this commit we relied on the notary generating a snapshot
for us. This is a bit problematic, as we use the same sign keys for
multiple notary repositories. Thus is was possible to replace the
targets section of a repository with a targets section of a
different repository that uses the same sign key.

We now calculate and sign the snapshot, thereby binding the
targets section data to a specific root.