Fix permission check in BsRequest.approval_handling and testcase
If I get the RequestControllerTest.test_create_request_approve_and_review
testcase right, a "/request/<id>?cmd=approve" http request is only allowed
if the currently logged in user has sufficient permissions to accept the
request <id>. However, the

    login_tom
    # try to approve without being empowered
    post "/request/#{id}?cmd=approve"
    assert_response 403

part of the old testcase only succeeds because the submit action's target
project "kde4" has the OBS:InitializeDevelPackage attribute set (the 403
is returned because "tom" has no permissions to initialize the source
package as a devel package). Hence, we use the "home:fred" project as
the submit action's target project. Changing this demonstrates that no
"approver" specific permission checks kick in (that is, the 403
assertion will fail because status 200 is returned). Hence, we
explicitly check in BsRequest.approval_handling if the currently logged
in user has sufficient permissions to accept the request.

A side effect of this change is that a user is only allowed to cancel
an existing approval of a request (cmd=cancelapproval), if she has
sufficient permissions to accept the request.