Fix flickering popover for build results
The build result popover can become very long, eg. when there are
many unresolvable package dependencies.
When this happens the popover would overlap with the link it belongs to
and thus would start to flicker when hovering over it (switching from
hovering state and back in a loop).

By displaying the popover on the right side of the link we prevent this
from happening.

Fixes #6200