Require system gems in api-deps package
Rake and Rack need to get installed on the system and not
with the bundled_gems archive because we require them for passenger.
This caused that an update of one of these gems in the Gemfile did not trigger
an update of the system gem when updating the RPM package which causes
a crash of the whole application.

Therefore we require them now. There are only 2 gems so it is reasonable to update
the version in the spec file manually for now when we update the gem. The RPM package
build will fail if we don't do it and prevents us from deploying and breaking our instance.