Show staging workflow link in project tabs
It was hidden because it was not released yet and planned to deploy to
production. However, we switched to work on a dedicated branch until
this feature is done so we can get rid of this hack.
This also makes development easier as we don't need to use the links
from the browsers history all the time.