Remove untracked requests from the list of missing reviews
In the new workflow we have a staged request relation that can be used
to assign requests to a staging project.

In the old workflow reviews that belong to 'untracked requests'[*] were included
here. But this is no longer needed, as confirmed by @coolo.

[*] Untracked requests are requests that do not belong to a staged
request, but have an open review for a staging project (via the
'by_project' attribute).