Implement Staging::StagedRequests create & destroy
These actions will assign / unassign the request to
the provided staging project and branch / destroy the
package in the staging project.

Currently the branching / destroying will block the request,
we plan to move this to a background job in the future
when we decided on a way to provide users feedback about
the status of a job.

One limitation is that this currently only supports submit actions.
For delete actions we added a TODO for the future.