Remove duplicated method
This method, and the commit it belongs to, appeared twice after
merging master into the staging-workflow branch.

The first time this commit got added was with the previous master
merge (#6250).

There was a small difference between both methods. I removed the one
that differed from the one we have in master.