Use --auto-gen-config to generate .rubocop_todo.yml
The cops in the Metrics department by default get `Max` parameters
generated in .rubocop_todo.yml, so that no offenses are reported the
next time you run rubocop. The new `--auto-gen-only-exclude` command
line option allow us to exclude the files instead. It will still change
the maximum if the number of excluded files is higher than the exclude
limit. This is specially useful when having only few files which
increase the maximum considerably, as it happens in our case for many cops.