Implement Project#monitor page in bootstrap
The project build montior table is now a DataTable and
all filter / search is performed in the DataTable now
instead of on the Server.

The monitor action still allows server filtering by passing parameters.
We still need to support this because of existing links.
However, when we already filtered on server side, the rendered table would
only contain a the filtered list which causes the client search to not
work as expected.

For now, we just disable the client search in this case and show a button
to remove all server side filters. This does a simple roundtrip to the server
without parameters which enables us to activate the client search again.