[backend] always run services on expanded links
Services are run on checkin, which usually works with expanded links
(keeplink mode), so this should not make a difference. What's changed
is if a remoterun is triggered, where the current way of running the
service on the unexpanded link is a source of lots of confusion.

The new way to run on expanded links is way more useful. Lets see
how it works in practice.

Also, the link code was changed to always ignore service results
of the link target if the link itself contains a service result.
This was only done for branches before, it is now more consistent.