Make the build state colors independent from bootstrap's pre-defined colors
There is no connection between primary-, secondary- and other signal colors
bootstrap defines, e.g. that text-info provides a blue tone we want to use
for the 'scheduled' state is just coincident.

Decoupling this prevents us from side-effects when changing the primary
color or using a different type of blue for notifications.