Disable also transitions in tests
Instead of injecting things into chrome, capybara's
AnimationDisabler is a rack middle ware that patches
the HTML to include a script and a style section. This
way we don't need to execute javascript.

Note that this is a beta feature of capybara, but exists
since 3.2.0 release (June) - I only discovered it by
chance though :)

Fixes #6457