Remove path helpers from our breadcrumbs
This replace the usage of path helpers in our breadcrumbs with
checks for action and / or controller name.
This means that we don't have to worry anymore whether a @project
or @package variable, that was needed by one of the helpers, was
set.
In addition we stop having to looking up the rooting table only
to check if we are on a show or index page.