Remove client filter from Project monitor
and do it on the server again.
The client search is not sufficient in its current implementation
because to add a filter you first need to remove all filters
which takes a lot of time for big projects like Factory.

We will think about a hibrid solution to enable client and server
side filter but for now the easiest solution is to always search on
the server.