Cache Package#show file list
InfluxDB performance measuring showed that this view is a bottleneck on Package#show with
a mean render time of ~ 280 ms (bootstrap) and 160 ms (bentoo). Therefore we
introduce now collection caching.

We also need to introduce timeago_tag jquery gem to be able to cache the table
otherwise the timeago would be cached as well causing wrong results.