Limit results for autocompletion queries to 50
We have a couple of pages that offer auto completion of searches of
packages, group and alike.
Depending on the search string this can return a lot of results and as
a result freeze the user's browser. Because of that, and because a
huge list of proposed projects (or groups, or ...) won't help users
anyway, we limit the returned results to 50.

Fixes #6558