Remove OBS label from breadcrumbs
On each page the breadcrumb starts with 'openSUSE Build Service'.
This text is taking away quite some space from the breadcrumbs and it
is not adding much value. Users already know that they are looking at
the OBS when they reached this page.
In case they don't we still have the OBS logo in the navbar.

The link to the main page is now part of the 'home' icon that is
part of each breadcrumb.