Stop using the flag model to calculate the flag matrix
Calculating the icon state per column is very wasteful
as it does nr(repository)*nr(architecture)*15 SQL queries.

Replace it with one query for all flags, store it in a has
and calculate the state in ruby - this is one SQL query and
some hash lookups per icon.

This saves for openSUSE:Tools ~75% in render time (~30 seconds).

Co-authored-by: Ana María Martínez Gómez <ammartinez@suse.de>