Refactor removing user from group
Instead of generating html code for each modal that asks for removing
a user from a group only one modal is created. Arguments for the modal
are passed with "data-*" attributes.

Rename the modal and the partials from "remove" to "delete", for being
consistent with the naming of the previous modals and partials.