Improve efficience of bento PackageController#buildresults
Pre calculate in the controller if a build result is outdated (not in database
but still temporarily in backend) instead of in the view, generating and SQL
query per result.

The fix in the new UI implied changing code in the model which was also executed
in the old one. As the fix is so simple now, fix it in the old UI as well.