Refactor delete of repository
Instead of generating html code for each modal that asks for deleting a
repository, only one modal is created. Arguments for the modal are passed
with "data-*" attributes.

The refactor has been done for repository cards and for download on
demand cards, as they use both the same modal.

Tests has been adapted.

The footer in the card has also been fixed. It was not well indented in
HAML.