Refactor request of deleting a repository
Instead of generating html code for each modal that asks for creating a
request of deleting a repository, only one modal is created. Arguments
for the modal are passed with "data-*" attributes.

The refactor has been done for repository cards and for download on
demand cards, as they use both the same modal.

The footer in the card has also been fixed. It was not well indented in
HAML.