Set nobody when request is processed
When we authenticate a user, we set User.current = .
Howerver, after the request is processed we never set it to nil.
If the next request does not require authentication and skips the
extract_user before filter, we still have the user from the previous
request set in User.current.
We should clean up what we set in Thread.current.