Provide a method to copy staging projects and their associations
We want to be able to use existing staging projects as template for
creating new ones. For that to work we need to be able copy
staging projects and a subset of their associations (flags,
relationships and repositories).
The deep_cloneable gem provides this functionality, which is why we use it here.

To be able to deep copy download repositories we had to update the arch
validation so that it works with associated data that is not yet saved.