Add reviewer group to the main project
When the main project has a reviewer group associated and a request
is sent to that project, a new review is created automatically. The problem
comes if the project has no reviewer group. To solve this, when a
staging workflow is created, we set its group to the project with
«reviewer» role.

Co-authored-by: David Kang <dkang@suse.com>