Improve efficiency of groups/users in staging workflow
Cache groups and users in the controller with two queries
instead of one query every time the group/user is displayed.

As we only have 32 groups in production and we don't expect this number to
increase or to be bigger in other instances of OBS, for the Groups we just
cache all of them instead of waisting time on finding out what of them we need.

Co-authored-by: Ana María Martínez Gómez <ammartinez@suse.de>