[WIP] Add report model
Copying staging projects can take longer amount of time, which is why
we run the code via delayed jobs.
To be able track the result of such asynchronous calls
(errors or information about copied data) and show themin the UI we add
a report model that can be attached to a staging project (the source we
copy from) or any other object via a polymorphic association.

An alternative would be to use the existing event system. But events
currently store associated objects in their payload which makes it
expensive to query them. Though we could change that.

Co-authored-by: Dany Marcoux <dmarcoux@suse.de>