Drop airbrake/errbit in favour of sentry
We are going to switch from errbit to sentry on our refference server because
sentry is a tad more sophisticated. This also means you can use this now with a
hosted version on https://sentry.io.

We do not keep errbit around in the bundle because of the size. Goodbye errbit,
you have been a true friend, good compagnion and helped us to track countless
errors! 💚