Reset backend location in RememberLocation module
because we now started to cache in the Backend API which means when
the backend is not called but fetched from cache, we never reset the
backend location. Now we always reset it after calling the method
which also is the right location to do so. This also wraps it now
into an ensure block to make it always gets cleared.