Refactor creation of request header
The header of the request diff header that we show in the UI depends on
the request action type. Since this involves quite some logic we move
the code out of the view.
Unfortunately we don't have the actual request action objects, but some
hash for this controller / view. Therefor we move the code into a
helper.