Disable view file links when we can not determine the previous revision
Each diff includes a link to a full view of the diff file.
In case of files that got deleted in the current revision,
this link refers to the previous revision.
In case of submit request diffs we only have a hashed revision.
Therefore we can not provide links to deleted files.

This also reduces the number of locals we hand over to each partial,
since we actually just need a package file path.