Move js event handler for comments to another partial
The 'comment/_content' partial is rendered multiple times. Therefore
the handler would be registered again and again.
By moving it to the comment show view and wrapping it into a document
ready statement we ensure that it's called only once, and for any
comment img element in the DOM.

Follow up of c2ec9cc2898df3ad329e93caa3241c9895eb96ad.