Remove unused spinner from the document layout
We previously used the 'invisible' class to hide the
user tab spinners. This caused the spinners to be hidden, but
still be rendered as part of the document.
Since these spinners also had the 'fa-spin' class, they kept
rotating while invisible and therefore caused the object style
to be recalculated every 200ms.

A sideeffect of removing the spinner icon was that the checkbox was
shifted one table column to the left. To fix this as well we had to
restructure the html code and the custom-control-* classes,
e.g. custom-control-label was assigned to a span element instead of
the label.