Hide monitor table while loading it
To avoid the overflow until the DataTable with the scroll is loaded.

This is needed because loading the DataTable is slow, as it is creating two
tables and moving the content around. Another alternative could be to stop
using the scroll from DataTables and just add it to the table, but then the
package column would move as well when scrolling what makes the table unusable.
In this case, and taking into account that the page is only big (= slow) in few
cases, makes sense to prioritize ease of use over performance. The user waits a
bit, but at least he gets a nice table.

The double table that DataTable renders is also causing a bug (which disappears
when not using the scroll from DataTable).