Make comment creation and deletion work when there are multiple comment sections
The javascript code was previously bound (not the js bind) to an html
id. Because of that we could only have one list of comments in the DOM.

With this change we can have multiple comment sections without having
them interfere with each other.
This is needed for showing comments of superseded requests in the request
view.