Make collapsible for history elements generic
Converts the collapsible text of history elements to a generic
helper that provides a simple way to make a text collapsible.
Collapsible in this case means that we show a substring of a
text that exceeds a given threshold and provide an 'expand' link.

The "collapsibility" can be applied by using the 'collapsible_text'
partial or by applying the correct classes manually (see our pattern library).