Hide decision box when there is nothing to do for the logged in user
Instead of showing a message that 'there's nothing to be done right now'
we hide the decision box.
This prevents reviewers with open reviews from making detours by having
to click on the "reviews" tab. The request decision can still be viewed
in the request history, like it was before.

Fixes #7026