Don't let disabled projects/repos pass stagings as acceptable
When osc crashes while selecting, sometimes we don't reenable the
staging project and then it becomes acceptable - which is an error.

We discussed this in the release manager round and there is no
valid use case of having disabled repositories *and* wanting it
to be acceptable. Mapping this to 'building' is obviously a little
fake, but unacceptable sounds too harsh too. So in building stagings
the RM will look from time to time to see what's taking so long and
she would notice then.