Align behavior of popovers throughout all views
Refactor initialization of popovers to remove duplication and align
behavior of all popovers. In a group of popovers, which is initialized
with a CSS selector, only one popover at a time can be shown. All shown
popovers are removed when a popover group is initialized.

Ignore JSHint errors regarding 'initializePopovers' is not defined.

Fixes #6950