[backend] BSSrcdiff: speed improvements
With the last commits we expanded .obscpio archives even
if both archives had the identical file path (i.e. the
package names are the same).
We now have a extra option in the fn() callback that makes
it return the unexpanded file name. We use this to
skip expansion if the names are identical.

This commit also puts the package name into the tmp file
used for .obscpio expansion, elimitating yet another
md5 aliasing problem.