Add workaround for rss notication events
With 394769ab1230020cd59c743a970a03ba07a2ce38 we changed the
format of the event payload. While we migrated all comment events
that were relying on them, we overlooked that the rss notifications
create temporary events based on their own event_payload.
Until we migrate them we need to add this workaround to prevent
exceptions when RSS Feeds that contain comment events are requested.