Set callbacks for staging project on project model
There where some callbacks specifically defined for staging projects
that have been moved to the project model. So they can be called even
if we perform some actions treating a staging project as regular
project.

For example, we can now delete a staging project via API using the same
endpoint that we use for any other kind of project and the right callbacks
will be triggered.

Also updates apidocs.