Don't crash on default reviews
Reviews added by users have a creator and a reason - reviews
added by project/package 'reviewer' role do not. So don't try
to show users if there are none.

This reverts quite big parts of #7128 because you don't need
an extra function to check the review reason - it's in the model.
Unfortunately we're not using the review model, but a hash
from it

Fixes #7164