Don't treat AttribNamespace and AttribType as static
We seed them, but we also change them during tests, so they
are not static. As such they should be truncated after the
tests and as such we need to create those DB entries we
need before the tests that need them.

As a *lot* of tests need them, load them before all of them.