Fix active review tabs when there is no decision box
When there is no decision box for the request (the user can not
accept or decline it) we have to mark the first review as
active (= visible).
Since we did not consider the index of the review list, all reviews
were visible when no the request decision box is hidden.

Fixes #7229