Re-design member list page(s) for groups
This was needed to be able show and edit the maintainer rights of users.
The group user edit and show page are now merged, which also solves the
different appearance of group users on both pages.

The "Actions" column in the group index view became superfluous because
it only contained a link to the, now removed, user edit view.
A link to the show page is still there.

Users can now see who is a maintainer of a group. Group maintainers can
change maintainer rights of other users.

Kudos for the appearance of the "group cards" goes to @dmarcoux