Migrate maintenance incident index view to new bootstrap UI
In the old UI we created a tooltip per incident that would contain the
project's build results. The content of the tooltip was loaded via
asynchronously. But since this was rarely used, we dropped it.
The build results can still be visited via a link.

This also adds a pagination option which is set to 'All' by default.
That's because the main users of this page usually want to see everything.