[backend] linkinfo_addtarget: fix corner case with project links
If there is a link/frozen link to an expanded link in an
inherited package we do not know if the link was expanded in
the origin project or not. This is relevant both for the lsrcmd5
(which might not be available yet) or the linkinfo's srcmd5.

So test where the trees are available and decide which project
to use.

This is a hack, but the information is simply not available so
we need to do some heuristic thing.