Catch unhandled backend errors and show a proper error message in the UI
This will catch Backend::Error and Timeout::Error exceptions and
redirect users the previous page. The exception will be send to errbit.

Many of the webui controllers call store without worrying about
possible exceptions that could be raised. This change makes sure
that we have a generic solution for this.