Align issues in load_from_xml and to_xml for Patchinfo
to_xml didn't have the issue URL. The fields were also not in the same
order in these 2 methods. This caused problems when updating a patchinfo
failed, which then rendered the edit view again, but with the missing
URL and fields in a different order. This resulted in issues not being
correctly rendered.

This happens because the edit action is initializing the patchinfo with
load_from_xml while the update action takes the data from the submitted
form without calling load_from_xml (to not override what the user
entered when validation errors happen). This wouldn't be an issue with
ActiveRecord.