Use server side pagination for the maintenance incident table
Most of the columns render data that is not part of the
project table. In case of the patchinfo, which is used in 3 columns,
we don't even have a database table we could query.
Because of these limitations only the 'Summary' column is sortable
right now.
For the same reason the search field was dropped.