Add a 'merging' state to project model
So far we were using the delayed jobs state for checking whether a
staging project is in "merging" state. But this has some drawbacks.
For example calling perform_now would start the project merge without
creating a delayed jobs instance.

This commit will introduce a 'merging' state that is used now to
reliably query the current state, which solves the mentioned issues.
In the far future we might even use a state machine for the
'overall_state' of a staging project. But this requires more drastic
changes and seem to be out of scope right now.

In addition we put the whole 'merging' code into a transaction to
avoid that a staging project ends up in a half-merged state.