Merge the release targets and build column of the incident table
Both columns are related to the release target. So it makes sense to
combine them. This also ensures that both targets are presented
together. Before, when there were multiple release targets and some
of them were wrapped to the next line, the release target name and
build icon were not on the same line.