Prevent writing of attributes if it's container is marked for destruction
After deletion of attribute instances we send a request to the backend
to update it there. But if the attribute deletion happens as part of the
project / package deletion we don't want this to happen.
Otherwise it could happen that the attribute deletion (after_commit
callback) is requested from the backend after the project / package
was deleted.

Admittedly this is a shot in the dark, as I couldn't reproduce this
issue. But this seems to be the only possible cause for the reported
failure.

Fixes #7419