Use create! instead of create in Update User#create_home_project
Make sure to raise an exception early on, if there is a problem.
Before this was failing because "You cannot call create unless
the parent is saved". But the actual problem was that the project
name was invalid.
This makes it a bit more obvious and easier to catch what the actual
problem is.

Found while debugging #7564.