Fix user home creation for users with certain login names
It's possible that user login names start with '.', ':' or '_'. This
causes a problem when creating home projects, because they can not
contain a ':' followed by one of aboves characters, like 'home:.tux'
in case of user login '.tux'.

On the other hand we support authentication via LDAP, Kerberos and alike
and we have some older user records start with a '_' in some OBS
instances.
Therefor we keep above user records valid and add a prefix for home
projects of such users, e.g. 'home:project.tux'.

Fixes #7564