Fix bs_request_action spec using User.current and increase coverage
After the change introduced by making User.current private and working
with User.session, User.session! instead, we forgot to refresh our
cassettes, which unsurprisingly, are failing when vcr is disabled.

The BsRequestAction model has in general a low spec coverage, so
extended it adding some missing specs

We disabled the vcr where it isnt necessary and deleted unecessary
cassettes

Co-authored-by: Henne Vogelsang <hvogel@opensuse.org>