Use content_for with running ruby code
As the `content_for` helper is not inserting any content directly in the
call, use '-' instead of '=' for calling the helper.

Also remove the parenthesis on all the calls of `content_for` with the
`:ready_function` parameter, as it is more often used without parenthesis
in the rest of the code.

And also remove an unneeded `:plain` sentence for a javascript
one-liner.