Fix bug on Request bugowner change link
In the Search Owner page there is a link to request a new bugowner for
the package. The new bugowner can be a user or a group, not both.
However, the action was always requiring both fields by mistake. Now only
one is required and sent.

Fixed in both old and new UI.

Reuses the same JavaScript function used to add a reviewer to a requet,
which has been slightly improved.

Co-authored-by: David Kang <dkang@suse.com>