[dist] Fixing our spec file license tag.
OBS source can be either used under GPL version 2.0 _or_ under
GPL version 3.0 terms.

This is not a re-licensing, just a fix of the former tagging in spec
files. (GPL v2.0 only and v3.0 only is basically not fullfillable, due
to the fact that v2.0 and v3.0 are incompatible)

Checked with SUSE legal dept.