Remove `height: auto` from CodeMirror Bootstrap theme
The height of CodeMirror instances are recalculated whenever a new
instance is initialized in the same container. This obviously blow ups
exponentially, making any page with multiple CodeMirror instances quite
slow to load.

The file was renamed and a comment was added to better convey what this
is really about.

Fixes #7752

Co-authored-by: Eduardo Navarro <enavarro@suse.com>