Fix comments word break, use hyphens instead
Comments were wrapped/broken on character boundaries instead of word
boundaries, making them hard to read. Caused by the CSS word break rule.

Removing that CSS rule doesn't help because the text exceeds the area.
For this reason, adding a hyphen to the long words seems to be a better
solution.