[ci] wait for outdated repos while testing build
Without this commit, the test cases in

0030_create_and_build_package.ts

will break, even if the result is outdated.

This is a regression that was introduced with #9b9f3a873d .
As we changed from "osc r" to "osc r -v" the output has changed,
but the regex to detect outdated repos was never fixed

With this fix we ensure, not to check as long as we have
outdated results.