Fix saving of project meta if containing current user as maintainer
If the uploaded XML contains the same relationship that the
source controller wants to create (current user as maintainer),
we get an invalid project and active record aborts the save!.

It took quite a while to debug this as we expected save! to throw
an exception for invalid projects. As this behaviour started
with rails 5.2.1 to trigger, we believe it is a Rails bug.

Fixes #7844

Co-authored-by: Ana María Martínez Gómez <anamaria@martinezgomez.name>