Don't shorten request source path text in webui2
Looking at the shortend/trimmed source path of a request
makes it sometimes difficult to identify the source, especially
if its through an interconnect (since its not a link to the source
in this case). Bootstrap handles long text quite well in this case.
So trimming the text after 40 characters is not necessary.