[dist] fix for upgrade obs-api 2.9 to 2.10
From 2.9 to 2.10 obsapidelayed.service changed to
obs-api-support.target. The restart and enabling of obs-api-support.target
was tried in the post script. This is to late as the obsapidelayed.service
is already deinstalled at this point and the systemctl command will return
an error.

This change ensures that "systemctl is-enabled/is-active" will be executed
in the %pre section while upgrading the package and the state will be stored
files in the %_run_dir. In the %post section we check for this files and
try to enable/start obs-api-support.target.

NEEDS BACKPORT TO 2.10