[dist] Include backend data dirs in package
Setting proper permissions for the top level directories in /srv/obs.
This was formerly done by "obsstoragesetup" only.
As "obsstoragesetup" should not be mandatory to run to get an working OBS
server, now this also done in the package. We also keep the old method in
"obsstoragesetup" as it gives the possibility to change OBS_BASE_DIR to users
needs.

Fixes #7984