Use ActiveRecord's #update_column instead of #update_attribute
For some changes we need to run the bootstrap specs with the VCR on. Doing so,
we get a ton of unwanted cassettes of requests going to the `/configuration`
endpoint in the backend.

We tried to fix that by using `#update_attribute` instead of the #update method
on the user factory so we didn't trigger the callbacks, but `#update_attributes`
skips the validations but not the callbacks. `#update_column` skips both the
validations and the callbacks, so its the one we have to use here.