Show the error raised when editting status comment
In the status tab/page of a project there is a list of failing packages.
A maintainer user can add a comment to each of the failing packages.
However, if there were any trouble while adding the comment, the error
wasn't visible.

Now the error is visible inside the comment cell.