Use coderay instead codemirror to present read-only diffs
To offer an options to provide to the users a way to check set of
one or more diffs, we were using CodeMirror, a javascript based Editor.

However when the request has more than a few diffs, it starts to get
slow. As drop-in solution we are going with CodeRay.

Fixes #8072