Enable RFS in bootstrap for responsive font-sizes
Improve the readability of the webui by scaling
the font-sizes depending on the screen size.
The proportions between headings and regular text will be
more harmonic on smaller screens.

Based on research the rfs scaling factor is set
to 2.5 in order to be close to the "major second"
scale, on small screens (360px width).
The rfs-base-font-size is set to 1rem (16px) to
not scale further down, since it's the recommended
font-size for regular text across all devices and screen sizes
to keep the readability.