Add required and hint into label element
Required asterisks and hints now are part of the labels. This improves
grouping all the information of an input field.
The hints were moved above the input boxes. This will help the user to
know the requirements before they start to add content.
To improve legibility the class `text-muted` was removed in the
hints, and `form-text` class was added.