Adapt maintenance workflow feature specs to Bootstrap
Some buttons and the flash message ID were slightly changed in the
Bootstrap UI.

"fill_in" calls were changed to match on the field IDs. Matching on the
labels' text didn't work anymore since they were redesigned to mark
fields as mandatory with an asterisk and a note.

Checkboxes are clicked with their label as we always do in the Bootstrap
UI.

Steps 6 and 7 cannot work since it's missing a functionality which was
available in Bento. See the issue #8207 on GitHub.