Move project_bread_crumb helper to obs_factory
Now that we are only using the new web interface, we want to get rid
of all the helpers only used with Bento.

We still use this helper, only in obs_factory, so we isolate it. This
way we make sure the application is not using this helper, except the
obs_factory part.

An exclude line in the RuboCop TODO file was also added, as now
instance variables are used in this helper.