Remove Bento actions from interconnects controller
Now that we are only using the new web interface, we want to adjust the
controllers, and get rid of the controller actions only used with Bento
for interconnects.

Move new UI views and partials to their definitive location, and
get rid of the Bento views and partials.

Remove what was only used in Bento controller:
- @interconnect instance variable.
- default_vaules private method.

Co-authored-by: David Kang <dkang@suse.com>