Replace Bento views by Bootstrap views for configuration
Now that we are only using the new web interface, we move new UI
views and partials to their definitive location, and get rid of
the Bento views and partials.

Adjust route of configuration breadcrumbs partial in interconnects.
This is required after moving the configuration breadcrumbs partial from
webui2/webui/ to webui/