Remove unused Bento add_reviewer_dialog action
Now that we are only using the new web interface, we get rid
of the Bento code and methods in the request controller.

This action was never called from a link with a route path helper, and
was substituted by a '#'.

Also remove the route, and one line in the spider tests.