Make the spider test crawl the sitemap (as good bots do)
We no longer have a way to list all projects nor a way to list
all packages in projects without having javascript due to use
of ajax. This made the spider test a game of luck and certain
packages were only found if they were listed in requests or
rebuild graphs (as it happened in Eduardo's PR for some reason).

But to have reproducible result, it's important to crawl all
pages to find also those we don't normally test in rspec