Restrict eager loading for HistoryElement::RequestReviewAdded class
The eager loading in the `request_history_elements` will only work for
those subclasses of HistoryElement::Request that implement the
`:review` association. In every other case, the association fails.

It does not make sense to implement that association in every
subclass to just make it pass, it would generate the wrong queries
because the meaning of `description_extension` is specific for every
subclass.

But restricting the association to only those subclasses that have it
defined will work.

Co-authored-by: Dany Marcoux <dmarcoux@suse.com>